#51 Sonar fixes for "When the "Rule-of-Zero" is not applicable, the "Rule-of-Five" should be followed" and "Floating point numbers should not be tested for equality"

+++ dynawo/sources/Modeler/DataInterface/DYNSwitchInterfaceIIDM.cpp
@@ -80,7 +80,7 @@ SwitchInterfaceIIDM::getComponentVarIndex(const std::string& varName) const {
 
 void
 SwitchInterfaceIIDM::exportStateVariablesUnitComponent() {
-  double state = getValue<int>(VAR_STATE);
+  int state = getValue<int>(VAR_STATE);
   if (state == OPEN)
     switchIIDM_.open();
   else
